필요 없는 소스 삭제
