+++ .idea/misc.xml
@@ -13,6 +13,7 @@
         <entry key="..\:/random-chatting/app/src/main/res/layout/file_upload_activity.xml" value="0.1" />
         <entry key="..\:/random-chatting/app/src/main/res/layout/file_upload_list_activity.xml" value="0.17916666666666667" />
         <entry key="..\:/random-chatting/app/src/main/res/layout/item_slider.xml" value="0.1" />
+        <entry key="..\:/random-chatting/app/src/main/res/layout/list_activity.xml" value="0.1" />
         <entry key="..\:/random-chatting/app/src/main/res/layout/slide_sample.xml" value="0.1" />
         <entry key="..\:/random-chatting/app/src/main/res/layout/test.xml" value="0.1" />
         <entry key="..\:/random-chatting/app/src/main/res/layout/user_list_activity.xml" value="0.12" />
