+++ app/src/main/AndroidManifest.xml
@@ -34,11 +34,9 @@
             android:exported="true"
             android:label="@string/app_name">
         </activity>
-        <activity android:name=".ListActivity" />
         <activity android:name=".FileUploadActivity" />
         <activity android:name=".UserListActivity" />
-        <activity android:name=".SlideSample"/>
-        
+
     </application>
 
 </manifest>
