--- app/src/main/java/com/example/random_chatting/ListActivity.java
@@ -1,35 +1,0 @@
-package com.example.random_chatting;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-public class ListActivity extends Activity {
-    Button btnClose, btnMoveHome;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.list_activity);
-
-        btnClose = (Button) findViewById(R.id.btnClose);
-        btnMoveHome = (Button) findViewById(R.id.btnMoveHome);
-
-        btnClose.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
-        });
-
-        btnMoveHome.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intentHome = new Intent(ListActivity.this, MainActivity.class);
-                startActivity(intentHome);
-            }
-        });
-    }
-}
