--- app/src/main/java/com/example/random_chatting/Tree.java
@@ -1,23 +1,0 @@
-package com.example.random_chatting;
-
-public class Tree {
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    String id;            // id
-    String userName;     // 이름
-}
