+++ app/src/main/java/com/example/random_chatting/UserListActivity.java
@@ -2,8 +2,10 @@ package com.example.random_chatting;
 
 import android.app.Activity;
 import android.content.Context;
+import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
+import android.os.Handler;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
@@ -41,7 +43,10 @@ import retrofit2.Callback;
 import retrofit2.Response;
 
 public class UserListActivity extends Activity {
+    public static List<TaskDTO.findUserInformationOutputDTO> mainFindUserInfomationList = new ArrayList<>();
+
     private TextView tvName, tvGender, tvAge, tvPhoneNumber;
+    private Button btnNextUser;
 
     // 이미지 폴더 경로 참조
     private StorageReference storageRef;
@@ -50,13 +55,8 @@ public class UserListActivity extends Activity {
     private ViewPager2 sliderViewPager;
     private LinearLayout layoutIndicator;
 
-    private String[] images = new String[]{
-            "https://cdn.pixabay.com/photo/2019/12/26/10/44/horse-4720178_1280.jpg",
-            "https://cdn.pixabay.com/photo/2020/11/04/15/29/coffee-beans-5712780_1280.jpg",
-            "https://cdn.pixabay.com/photo/2020/03/08/21/41/landscape-4913841_1280.jpg",
-            "https://cdn.pixabay.com/photo/2020/09/02/18/03/girl-5539094_1280.jpg",
-            "https://cdn.pixabay.com/photo/2014/03/03/16/15/mosque-279015_1280.jpg"
-    };
+    //page (1페이지 부터 시작)
+    private Integer currentPageCnt = 1;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -67,101 +67,11 @@ public class UserListActivity extends Activity {
         tvGender = (TextView) findViewById(R.id.user_list_activity_tv_gender);
         tvAge = (TextView) findViewById(R.id.user_list_activity_tv_age);
         tvPhoneNumber = (TextView) findViewById(R.id.user_list_activity_tv_phone_number);
-
+        btnNextUser = (Button) findViewById(R.id.user_list_activity_btn_next_user);
         storageRef = FirebaseStorage.getInstance().getReference();
 
-        /**
-         * 정보 조회 시작
-         */
-        Call<String> call = Retrofit_client.getApiService().findUserInformation(
-                "select");
-
-        call.enqueue(new Callback<String>() {
-            //콜백 받는 부분
-            @Override
-            public void onResponse(Call<String> call, Response<String> response) {
-                String jsonResponse = response.body();
-                try {
-                    JSONObject jsonObject = new JSONObject(jsonResponse);
-                    if (jsonObject.optString("status").equals("true")) {
-                        //Toast.makeText(UserListActivity.this, "File Get Error!!2", Toast.LENGTH_SHORT).show();
-                        JSONArray jsonArray = jsonObject.getJSONArray("result");
-                        List<TaskDTO.findUserInformationOutputDTO> findUserInfomationList = new ArrayList<>();
-
-                        for (int i = 0; i < jsonArray.length(); i++) {
-                            try {
-                                List<String> fileNameList = new ArrayList<>();
-
-                                JSONObject loopJsonObject = jsonArray.getJSONObject(i);
-                                // Pulling items from the array
-                                TaskDTO.findUserInformationOutputDTO findUserInformationOutput = new TaskDTO.findUserInformationOutputDTO();
-                                String id = loopJsonObject.getString("id");
-                                String userName = loopJsonObject.getString("userName");
-                                String gender = loopJsonObject.getString("gender");
-                                String age = loopJsonObject.getString("age");
-                                String phoneNumber = loopJsonObject.getString("phoneNumber");
-                                String fileName0 = loopJsonObject.getString("fileName0");
-                                String fileName1 = loopJsonObject.getString("fileName1");
-                                String fileName2 = loopJsonObject.getString("fileName2");
-                                String fileName3 = loopJsonObject.getString("fileName3");
-                                String fileName4 = loopJsonObject.getString("fileName4");
-                                String fileName5 = loopJsonObject.getString("fileName5");
-                                String countIdx = loopJsonObject.getString("countIdx");
-
-                                //decode
-                                userName = URLDecoder.decode(userName, "utf-8");
-                                gender = URLDecoder.decode(gender, "utf-8");
-
-                                fileNameList.add(fileName0);
-                                fileNameList.add(fileName1);
-                                fileNameList.add(fileName2);
-                                fileNameList.add(fileName3);
-                                fileNameList.add(fileName4);
-                                fileNameList.add(fileName5);
-
-                                //공백 제거
-                                fileNameList.removeAll(Arrays.asList("", null));
-
-                                findUserInformationOutput.setId(id);
-                                findUserInformationOutput.setUserName(userName);
-                                findUserInformationOutput.setGender(gender);
-                                findUserInformationOutput.setAge(age);
-                                findUserInformationOutput.setPhoneNumber(phoneNumber);
-                                findUserInformationOutput.setFileNameList(fileNameList);
-                                findUserInformationOutput.setCountIdx(countIdx);
-
-                                findUserInfomationList.add(findUserInformationOutput);
-                            } catch (JSONException e) {
-                                // json catch
-                            } catch (UnsupportedEncodingException e) {
-                                e.printStackTrace();    // encoding catch
-                            }
-                        }
-
-                        //random 처리
-                        Collections.shuffle(findUserInfomationList);
-
-                        //정보 표시 (비동기 처리로 수정 해야함.)
-                        showInformation(findUserInfomationList.get(0));
-
-                        Toast.makeText(UserListActivity.this, "조회 성공", Toast.LENGTH_LONG).show();
-                    } else {
-                        Toast.makeText(UserListActivity.this, "조회 실패 : " + jsonObject.optString("query"), Toast.LENGTH_LONG).show();
-                    }
-                } catch (JSONException e) {
-                    Toast.makeText(UserListActivity.this, "조회 실패 (catch) : " + e.getMessage().toString(), Toast.LENGTH_LONG).show();
-                }
-            }
-
-            @Override
-            public void onFailure(Call<String> call, Throwable t) {
-                Toast.makeText(UserListActivity.this, "조회 실패 (onFailure)", Toast.LENGTH_LONG).show();
-            }
-        });
-
-        /**
-         * 정보 조회 끝
-         */
+        FindUserInformationTaskRxJava findUserInformationTaskRxJava = new FindUserInformationTaskRxJava(this);
+        findUserInformationTaskRxJava.runFunc();
 
         /* 전체 불러오기 주석처리(현재 안씀)
         btDirectoryImageLoad.setOnClickListener(new View.OnClickListener() {
@@ -172,10 +82,23 @@ public class UserListActivity extends Activity {
         });
          */
 
+
+        btnNextUser.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if(currentPageCnt < mainFindUserInfomationList.size()) {
+                    showInformation(mainFindUserInfomationList.get(currentPageCnt));
+                    currentPageCnt++;
+                }else{
+                    Toast.makeText(getApplicationContext(), "더 이상 회원이 없습니다.", Toast.LENGTH_SHORT).show();
+                }
+            }
+        });
     }
 
     //slider
     private void setupIndicators(int count) {
+        layoutIndicator.removeAllViews();
         ImageView[] indicators = new ImageView[count];
         LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(
                 ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
@@ -209,9 +132,8 @@ public class UserListActivity extends Activity {
             }
         }
     }
-    //
 
-    private void showInformation(TaskDTO.findUserInformationOutputDTO info) {
+    public void showInformation(TaskDTO.findUserInformationOutputDTO info) {
         tvName.setText(info.getUserName());
         tvGender.setText(info.getGender());
         tvAge.setText(info.getAge());
@@ -227,8 +149,7 @@ public class UserListActivity extends Activity {
     }
 
     //이미지 slide 추가 해야함.
-    private void showPhoto(Context context, String[] fileNameArray) {
-
+    public void showPhoto(Context context, String[] fileNameArray) {
         sliderViewPager = findViewById(R.id.sliderViewPager);
         layoutIndicator = findViewById(R.id.layoutIndicators);
 
@@ -244,7 +165,6 @@ public class UserListActivity extends Activity {
         });
 
         setupIndicators(fileNameArray.length);
-
     }
 
     //폴더에 있는 이미지 전부 가져와서 동적으로 뿌려준다.
