+++ app/src/main/res/layout/activity_main.xml
@@ -107,29 +107,8 @@
         android:id="@+id/activity_main_btn_upload_list"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="174dp"
-        android:text="파일 업로드 목록"
-        app:layout_constraintLeft_toRightOf="@+id/activity_main_btn_save"
-        app:layout_constraintStart_toStartOf="@+id/nav_host_fragment_activity_main"
-        app:layout_constraintTop_toBottomOf="@+id/activity_main_et_phone_number" />
-
-    <Button
-        android:id="@+id/activity_main_btn_list"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="87dp"
-        android:text="회원 목록"
-        app:layout_constraintLeft_toRightOf="@+id/activity_main_btn_save"
-        app:layout_constraintStart_toStartOf="@+id/nav_host_fragment_activity_main"
-        app:layout_constraintTop_toBottomOf="@+id/activity_main_et_phone_number" />
-
-    <Button
-        android:id="@+id/activity_main_slide_sample"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="48dp"
-        android:text="테스트"
+        android:layout_marginStart="90dp"
+        android:text="목 록"
         app:layout_constraintLeft_toRightOf="@+id/activity_main_btn_save"
         app:layout_constraintStart_toStartOf="@+id/nav_host_fragment_activity_main"
         app:layout_constraintTop_toBottomOf="@+id/activity_main_et_phone_number" />
