--- app/src/main/res/layout/slide_sample.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/sliderViewPager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <View
-        android:id="@+id/viewFadingEdge"
-        android:layout_width="match_parent"
-        android:layout_height="55dp"
-        android:background="@drawable/bg_fading_edge"
-        app:layout_constraintBottom_toBottomOf="@id/sliderViewPager" />
-
-    <LinearLayout
-        android:id="@+id/layoutIndicators"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:gravity="center"
-        android:paddingBottom="8dp"
-        app:layout_constraintBottom_toBottomOf="@id/viewFadingEdge"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
