+++ app/src/main/res/layout/user_list_activity.xml
@@ -93,6 +93,7 @@
             app:layout_constraintTop_toBottomOf="@+id/user_list_activity_tv_gender" />
 
         <TextView
+            android:id="@+id/user_list_activity_tv_label_phone_number"
             android:layout_width="100dp"
             android:layout_height="wrap_content"
             android:text="핸드폰 번호 : "
@@ -106,6 +107,24 @@
             android:layout_marginStart="100dp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/user_list_activity_tv_lable_age" />
+
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/user_list_btn_layout"
+        android:layout_width="match_parent"
+        android:layout_height="100dp"
+        android:gravity="left"
+        app:layout_constraintTop_toBottomOf="@+id/maskImageLayout">
+    <Button
+        android:id="@+id/user_list_activity_btn_next_user"
+        android:layout_width="200dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="100dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:text="다  음"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
